Structure folder refactored
