+++ src/components/common/includes/NavBar.vue
