+++ src/components/posts/mixins/list.js
